Improve camera picture quality
The camera used in our setup is the NoIR cam which has the issue
of having a strong pink tint in images taking with daylight. The
camera firmware has a special white balance mode "greyworld" to
compensate for this. However, this mode is not yet available
through the picamera library. Therefore, following a suggestion
from this SO comment:
https://github.com/raspberrypi/firmware/issues/1167#issuecomment-511798033
we're adding the new white balance mode to the camera.

Furthermore, also change the camera's resolution to a format that
makes use of the entire sensor (scaled down to 25%). Also play
with the brightness and contrast parameters to get the best image
possible.